+++ sys/ofed/drivers/infiniband/core/ib_sysfs.c
@@ -874,7 +874,7 @@ static ssize_t set_stats_lifespan(struct kobject *kobj,
 	struct hw_stats_attribute *hsa;
 	struct rdma_hw_stats *stats;
 	int msecs;
-	unsigned long jiffies;
+	int jiffies;
 	int ret;
 
 	ret = kstrtoint(buf, 10, &msecs);
